+++ widgets/mixins/__init__.py
@@ -1,0 +1,16 @@
+"""
+Module: mixins/__init__.py
+
+Author: Michael Economou
+Date: 2025-12-04
+
+Mixins for reusable widget behavior.
+"""
+
+from widgets.mixins.selection_mixin import SelectionMixin
+from widgets.mixins.drag_drop_mixin import DragDropMixin
+
+__all__ = [
+    "SelectionMixin",
+    "DragDropMixin",
+]